added check jar exissts chjeck to to delete message command

+++ src/main/java/com/tikoJar/DTO/QueryHandler.java
@@ -115,7 +115,9 @@ public class QueryHandler {
     public void deleteMessage(boolean includedMessageID, String messageID){
         boolean messageDeleted = true;
         if(includedMessageID){
-            deleteMessageQuery(messageID);
+            if(checkIfJarExists()){
+                deleteMessageQuery(messageID);
+            }
         }
         responseBuilder.deleteMessageResponse(includedMessageID, messageDeleted);
     }
